修改站点模块设置错误

+++ app/Http/Controllers/SiteModuleController.php
@@ -2,9 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\StoreSiteModuleRequest;
 use App\Http\Resources\ModuleResource;
 use App\Models\Module;
 use App\Models\Site;
+use App\Models\SiteModule;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
 
@@ -17,20 +19,25 @@ class SiteModuleController extends Controller
 
     public function index(Site $site)
     {
+        $this->authorize('viewAny', SiteModule::class);
         $modules = $site->modules()->latest()->paginate();
         return ModuleResource::collection($modules);
     }
 
     public function store(Request $request, Site $site)
     {
+        $this->authorize('create', SiteModule::class);
+
         $site->modules()->syncWithoutDetaching([request('mid')]);
         app('permission')->syncAllModulePermissions($site);
 
         return $this->success('模块添加成功');
     }
 
-    public function destroy(Site $site, Module $module)
+    public function destroy(Site $site, SiteModule $module)
     {
+        $this->authorize('delete', $module);
+
         $site->modules()->detach($module->id);
         $site->permissions()->where('module_id', $module->id)->delete();
         return $this->success('模块删除成功');
